fix: Update dark mode implementation to work with Next.js server components
