+++ app/layout.tsx
@@ -1,7 +1,8 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import './globals.css'
+import DarkModeProvider from './components/DarkModeProvider'
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -23,39 +24,12 @@ export default function RootLayout({
 }: Readonly<{
   children: React.ReactNode
 }>) {
-  // This effect will run on the client side to detect dark mode preference
-  useEffect(() => {
-    // Check if the user prefers dark mode
-    const prefersDarkMode = window.matchMedia('(prefers-color-scheme: dark)').matches
-
-    // Apply dark mode class to html element if user prefers dark mode
-    if (prefersDarkMode) {
-      document.documentElement.classList.add('dark')
-    } else {
-      document.documentElement.classList.remove('dark')
-    }
-
-    // Listen for changes in color scheme preference
-    const mediaQuery = window.matchMedia('(prefers-color-scheme: dark)')
-    const handleChange = (e: MediaQueryListEvent) => {
-      if (e.matches) {
-        document.documentElement.classList.add('dark')
-      } else {
-        document.documentElement.classList.remove('dark')
-      }
-    }
-
-    mediaQuery.addEventListener('change', handleChange)
-
-    // Clean up event listener on component unmount
-    return () => {
-      mediaQuery.removeEventListener('change', handleChange)
-    }
-  }, [])
-
   return (
     <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>{children}</body>
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
+        <DarkModeProvider />
+        {children}
+      </body>
     </html>
   )
 }
